feat(python): Support shortcut eval of common boolean filters in SQL interface "WHERE" clause
