+++ crates/polars-sql/src/context.rs
@@ -847,8 +847,28 @@ impl SQLContext {
         expr: &Option<SQLExpr>,
     ) -> PolarsResult<LazyFrame> {
         if let Some(expr) = expr {
-            let schema = Some(self.get_frame_schema(&mut lf)?);
-            let mut filter_expression = parse_sql_expr(expr, self, schema.as_deref())?;
+            let schema = self.get_frame_schema(&mut lf)?;
+
+            // shortcut filter evaluation if given expression is just TRUE or FALSE
+            let (all_true, all_false) = match expr {
+                SQLExpr::Value(SQLValue::Boolean(b)) => (*b, !*b),
+                SQLExpr::BinaryOp { left, op, right } => match (&**left, &**right, op) {
+                    (SQLExpr::Value(a), SQLExpr::Value(b), BinaryOperator::Eq) => (a == b, a != b),
+                    (SQLExpr::Value(a), SQLExpr::Value(b), BinaryOperator::NotEq) => {
+                        (a != b, a == b)
+                    },
+                    _ => (false, false),
+                },
+                _ => (false, false),
+            };
+            if all_true {
+                return Ok(lf);
+            } else if all_false {
+                return Ok(DataFrame::empty_with_schema(schema.as_ref()).lazy());
+            }
+
+            // ...otherwise parse and apply the filter as normal
+            let mut filter_expression = parse_sql_expr(expr, self, Some(schema).as_deref())?;
             if filter_expression.clone().meta().has_multiple_outputs() {
                 filter_expression = all_horizontal([filter_expression])?;
             }
