+++ crates/polars-sql/src/function_registry.rs
@@ -1,4 +1,4 @@
-//! This module defines the function registry and user defined functions.
+//! This module defines a FunctionRegistry for supported SQL functions and UDFs.
 
 use polars_error::{polars_bail, PolarsResult};
 use polars_plan::prelude::udf::UserDefinedFunction;
