+++ crates/polars-sql/src/lib.rs
@@ -7,6 +7,7 @@ mod functions;
 pub mod keywords;
 mod sql_expr;
 mod table_functions;
+mod types;
 
 pub use context::SQLContext;
 pub use sql_expr::sql_expr;
